[neog-48] feat: Added user details form for new user.
