+++ src/pages/user-details/UserDetails.jsx
@@ -1,0 +1,11 @@
+import EditProfileForm from "../../components/edit-profile-form/EditProfileForm";
+
+const UserDetails = () => {
+    return (
+        <div>
+            <EditProfileForm/>
+        </div>
+    );
+};
+
+export default UserDetails;
